final_project/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
jinja2>=3.1
itsdangerous==2.2.0
python-multipart==0.0.26
pandas==3.0.2
numpy==2.4.4
scikit-learn==1.8.0
sqlalchemy==2.0.49
psycopg[binary]==3.3.3
azure-storage-blob==12.27.0
httpx==0.28.1
pytest==9.0.3
